requirements.txt
fastapi
sqlalchemy
psycopg2-binary
uvicorn
pydantic[email]
pandas
openpyxl